Move tests to hardware/tests/tests_utils.py
